cloud_functions/completion/requirements.txt
google-auth>=1.3.0
google-cloud-firestore>=0.27.0
json-logic-qubit==0.9.1
algoliasearch==2.4.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
firebase-admin==4.5.3
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
httplib2==0.19.1
idna==2.10
msgpack==1.0.2
numpy==1.20.2
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.4
dnspython==1.16.0
pandas==1.2.3
pymongo==3.11.3
python-dateutil==2.8.1
